Move CsvEpochSource from example into main code
CsvEpochSource was originally implemented for the purpose of demonstrating how to create a subclass of WritableEpochSource. However, if `'name': 'animal_state'` is removed, it's fully general and could be very useful to anyone interested in saving epochs using CSVs. Consequently, it's a great candidate for adoption into the main code.

This change migrates the implementation of CsvEpochSource to the main code so that users do not need to implement this subclass themselves. They are still invited to create alternative implementations of subclasses of WritableEpochSource that use different file types. Dependence of ephyviewer on pandas is avoided by requiring pandas only when CsvEpochSource is instantiated.
